% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparePipeline.R
\name{preparePipeline}
\alias{preparePipeline}
\title{preparePipeline}
\usage{
preparePipeline(workdir = c(), what = "readFileOnly", brand = "AX", fileID = 1)
}
\arguments{
\item{workdir}{path to location where code can store test files and results}

\item{what}{character to specify what to profile readfile}

\item{brand}{character with name of brand: Axivity}

\item{fileID}{number to indicate which file to read}
}
\value{
no object is returned
}
\description{
preparePipeline
}
